+++ rmotr_todo/todo/views.py
@@ -1,22 +1,20 @@
 from django.shortcuts import render, redirect
-from django.http import HttpResponseBadRequest, HttpResponseNotFound
+from django.http import (HttpResponseBadRequest,
+                         HttpResponseNotFound,
+                         HttpResponseForbidden)
+from django.contrib.auth.decorators import login_required
 from todo.models import Item, List
 from todo.forms import ItemForm
 from django.core.exceptions import ObjectDoesNotExist
 
 
+@login_required
 def home(request):
-    return render(request, 'home.html', {'form': ItemForm()})
-
-
-def new_list(request):
-    form = ItemForm(data=request.POST)
-    if form.is_valid():
-        list_ = List.objects.create()
-        Item.objects.create(text=request.POST['text'], list=list_)
-        return redirect(list_)
-    else:
-        return render(request, 'home.html', {'form': form})
+    try:
+        list_ = List.objects.get(user=request.user)
+    except ObjectDoesNotExist:
+        list_ = List.objects.create(user=request.user)
+    return redirect(list_)
 
 
 def view_list(request, list_id):
@@ -24,6 +22,9 @@ def view_list(request, list_id):
         list_ = List.objects.get(id=list_id)
     except ObjectDoesNotExist:
         return HttpResponseNotFound()
+
+    if not list_.owned_by(request.user):
+        return HttpResponseForbidden()
 
     form = ItemForm()
 
@@ -34,12 +35,16 @@ def view_list(request, list_id):
             return(redirect(list_))
     return render(request, 'list.html', {'list': list_, 'form': form})
 
+
 def delete_item(request, list_id, item_id):
     try:
         list_ = List.objects.get(id=list_id)
         item = Item.objects.get(id=item_id)
     except ObjectDoesNotExist:
         return HttpResponseNotFound()
+    if not item.owned_by(request.user):
+        return HttpResponseForbidden()
     if request.method == 'POST':
         item.delete()
     return redirect(list_)
+
